created test file for animal management
